Added pagePermission function in meta.php
added pagePermission function in meta.php, this feature requires the argument $role, which takes an integer for the current role. the roles are displayed as comments within this function

This function is ment to be used as gatekeeper to keep unauthorized users away from the page. call the function at the top of every page!